refactor: 면접 대기방 페이지에서 스타일드 컴포넌트 별도 파일로 분리
- 스타일드 컴포넌트로 인해 코드 양이 많아짐에 따라,  가독성을 개선하고 비즈니스 로직과 구분하기 위해 별도 파일로 분리함.
